requirements.txt
aiohttp==3.8.5
asyncio==3.4.3
beautifulsoup4==4.12.2
feedparser==6.0.10
PyGithub==1.59.0
python-dotenv==1.0.0
requests==2.31.0
anthropic==0.5.0